Do not expire artificial water near unloaded area boundary
Sometimes sponge squeezers are built too close to a mapblock
boundary, and rarely fail due to the artificial water being
loaded but the lens powering the squeezer being unloaded or
inactive.  Try to detect these borderline cases and prevent the
water from disappearing.  It would be somewhat unfair to expect
players to predict and prevent this failure mode, especially in
certain high stakes play modes like skyblocks, so try to prevent
it if we can.