Sort discoveries chronologically, newest first
Note that if the exact wording of a hint changes, then those hints
may be pushed to the top of the list.  This is an internal
limitation and cannot be prevented without assigning some other
stable unique identifier to hints (a lot of work and probably not
worth it for now).

A tiny amount of timing jitter is applied to hint discover time
(<1ms) to ensure that hints are always in a stable order.

Hints are identified by a short hash of the hint text, so it is
also possible (but unlikely) to have collisions that cause
hints to sort earlier in order than they normally should.