Revert "Exact sprout selectionbox tweak"
This reverts commit 16ed60d43fdba76d01d5800e561bd0c401085399.

After playing with this for a while, it's too ugly and the sprout
bit itself isn't really tall enough to warrant a tiny extra
selectionbox.  I had done this because the catrealm mod has a
bush stem node that's bigger, and applying it to the bush sprout
was consistent, and thus the eggcorn sprout too ... but maybe
it makes more sense to just exempt the tiny sprouts.