Support teleporting to offline players
The YCTIWY mod is already keeping track of offline player
positions, so allow admins to use those for teleportation.  This
is sort of like the szutil_offlinepos mod, but as a standard
game feature now.  It can be useful for admins to quickly locate
players who are trapped and need rescue, or suspected of griefing
and need to be confirmed.