Immediate optic state acquisition on placement
When placing an optic node into an existing beam path, it
immediately acquires the correct state rather than being placed
as "inactive" and needing one tick to activate.  This would cause
a "down" pulse to travel down networks when placing a "vampiric
tap" into an existing beam path, even if it was placed with the
correct initial rotation.  This reduces a source of noise in
optic circuit building that doesn't appear to add any meaningful
challenge, only nuisance.