Mitigate optic duplication bug
Optic nodes are each assigned a "family" of nodes that
are equivalent things with different power states.  Optic
commit is not allowed to replace any node with another
node outside of the source node's family.  This at least
prevents optic_commit replacing an air node with an
optic node (the way the dupe manifests).  Log a warning
whenever this happens.

Unfortunately we still don't yet know the actual cause of
the problem, so this is just a workaround, not an actual
fix.