Fix broken stylus programming
...along with possibly other downstream bugs.

We had stripped the metadata off the wield item in the
pummel state tracking to make it so that changes to
wield item metadata (especially things that could be
changed by AISM while the player is using the thing)
would not cancel a pummel, but forgot that the wield
metadata that gets put into the pummel state is ALSO
fed to the recipe check.  Use a separate key for this
stripped only-for-comparison version, and keep the
proper itemstack with full meta for recipe checks.