Plan to rework the falling node / cave-in mechanic
Spontaneous node falling was intended to serve two
purposes originally:

- Prevent falling nodes from remaining nonsensically
  suspended indefinitely from bad mapgen.
- Add a bit of tension to the game, back when it was
  envisioned as more "hardcore" in more traditional
  a sense.

The feature no longer fits with the tone and themes
of the game today.  Sudden cave-ins caused by the
"sound/vibrations" of players disturbing an area are
largely indifferent to most players, confusing to
some, and very much hated by a few.