Simplify progressive zoom controls
Just hold the zoom button for a couple of seconds
(gives the player a chance to take screenshots with
the fixed zoom level) and then progressive zoom will
kick in automatically as long as zoom and no other
input control is held down.

This also prevents problems with the sneak/aux
causing unwanted movement while zooming.