Stop asking for artwork help
It's apparent that the barrier of needing to be able to maintain
a consistent art style myself without depending on external artists
is a significant barrier to being able to accept all but the most
isolated artwork changes.  I really only NEED occasional help with
specific things, like maintaining the player character model and
animations, and in other cases it's probably better for people to
just make texturepacks for NodeCore and distribute them themselves.