Force stuck item settling
If a "moving" item is not moving by at least 25cm per
second, then assume it's "stuck" and allow it to build
up "potential" for tunneling, up to a sane limit (to
prevent it from getting so large that it's likely to cover
mostly unloaded mapblocks, which would actually
reduce the probability of settling).

This should get rid of any remaining "deadlocked"
item entities that can happen in the normal course of
gameplay.