Fix tools acquiring a position offset in door dig animation
The animation was repositioning the tool so that when it returned,
it would return to the center of the node space instead of sitting
on the bottom, the way visinv ents normally are supposed to.

Instead of using node centers for the positioning in the animation,
preserve the offset relative to center in both key positions in
the animation.

Tools should now return to the same offset where they started.