Experimental workaround for failed teleports
If we (or any mod) attempts to teleport a player a
non-trivial amount of distance, then detect if the teleport
was "regressed" by weird client lag shenanigans, by seeing
if the player is closer to the source than the destination
in less time than it would take to actually travel that
distance under reasonable circumstances.

If we detect this has happened, then retry the teleport
and log a warning.