Remove hacky cache clear on player logout
I'm actually not confident we can reliably tell the difference
between a player logout and timeout.  It's also unnecessary since
we already have a 60 second upper bound; if a player ever gets
trapped in some kind of teleportation loop, they can just stay
offline for 60 seconds to work around it (and hopefully report the
problem).