Force texture scaling for molten glass and amalgamation
For a high-resolution texture pack to properly support these, it
MUST use overrides, either to replace the texture with a pre-baked
one, or to correct the combine layer offsets.  If a texture pack
does not do this, rescale the images down to 16x16 so that the
resultant image is at least properly aligned at some resolution.