Selective register_on_nodeupdate API
The old API simply registered functions, which listened for all
events and received all parameters.  This was inefficient because:

- Not all calls actually used the "node" parameter, so if we don't
  need it, and the event that triggers the notification doesn't
  provide one, we used to have to do a get_node for it.
- A lot of things that can change a node in the world can't
  actually cause certain things to change.  For example,
  remove_node can never cause a cookable node to be in a position
  for a recipe check (and making plain air cookable would not
  be sane).

Allowing us to selectively ignore certain events, and to indicate
that we aren't using the "node" param so the API doesn't need to go
out of its way to provide it, should make a lot of things more
efficient.

The biggest problems we were seeing were when a lot of liquids were
transforming due to "pumhell" deep underground, where pumice
freezing and melting causes constant reflows of pumwater.  This is
an intentional game mechanic to make those areas dangerous to
explore or try to develop without taming them, but the only notify
callback we really need to trigger is the optic revalidation one,
which should just be checking for indexed beams to reevaluate and
finding none in most cases.  We were seeing a ton of lag when in
these areas, leading to the hypothesis that it's all the other
unnecessary checks that causes this slowdown.