Don't re-defer falling nodes when unloaded
Nodes that are stuck in a deferred check state can get unloaded,
and we don't want to keep trying to re-check them again, even if
we are only doing it every few seconds now.  Allow them to fall
out of the queue entirely, and we can pick them up again next time
they get loaded.