Revert "Experimentally nerf falling node unsuspend LBM"
Testing did not reveal this to have a meaningful impact; most of
the lag I was hoping to prevent with it is still there.  Even though
this is a big operation and runs tens of thousands of iterations
per second in some cases, it's still efficient enough overall that
it doesn't impact lag significantly.

Leaving it disabled would cause there to be bugs/exploits with
tall columns of stuff remaining suspended in mid air if the upper
parts are unloaded when the bottom is dug out.  While there is no
way to avoid that entirely, as long as the middle area where
some falling node is floating above air gets loaded eventually,
it will at least self-resolve with the feature intact as
originally.

This reverts commit 580fcc462292f495b531735ca098faae9060b9c7.