Account for loaded but inactive sponge squeezers
Specifically for the case of sponge squeezers, not only do we need
enough map around the artificial water to be loaded, we need it to
actually be active, to be certain that the squeezer has an
opportunity to run and keep the water fresh.  This change should
prevent specific water nodes from disappearing from sponge
squeezers when they're right on the edge of the active mapblock
range.

This genericizes the API for checking the state of a mapblock, and
adds a function to check if a mapblock is active, not merely loaded.