Direct lode ore hints in mapgen
Instead of just generating "clouds" of "hint" nodes in the same
general strata where lode ore could also exist, have the hints
directly "radiate" outward from actual ore nodes.  This means that
finding a hint node should always guaranteed lead you to an ore
node, as long as you haven't already harvested it, and as long as
you guess the direction correctly.

One overall consequence of this is that the ore hints are a lot
sparser overall, and there are lots of areas where you're near
ore-bearing stone but can't tell because the hints just happened
not to reach that far.

This might be slightly expensive compared to the old method, since
we're essentially casting rays in Lua, but it seems to be
surprisingly efficient in testing so far.