Consolidate all table_shuffle reimplementations
I've apparently needed to Knuth-shuffle a table pretty often, and
have reimplemented it multiple times.  We should instead assume
that the "standard" table.shuffle extension that MT adds should be
reasonably "optimal" and just use that everywhere.  If it turns
out not to be optimal then it's still valuable to consolidate the
implementations.