Get rid of objects_inside_radius use
These are terribly inefficient, especially when there are a ton
of entities, as is the case when there are a lot of visenv ents.

For nodefall damage, filtering by whether the entity is a player
first is much more efficient than by position, since the rejection
ratio on larger servers is higher.

For checking if doors or totes are obstructed by falling or item
entities, this isn't even strictly necessary and maybe we should
just skip it and allow them to get "tunneled" in and settle
naturally.  At higher step sizes, it's questionable whether these
would even be handled accurately in the first place.