Soaking reform to reduce metadata writes
Instead of always pushing rate * time -> total, only commit this
when there has been a change of rate or override of total.  This
means that rather than updating metadata every time a soaking ABM
runs on a node, it will only get updated when "steady" soaking has
been interrupted, greatly reducing the volume of metadata writes
and dirty mapblocks.

This should work fine because the math was already designed to
work reasonably well when updates are "skipped" due to the map
not being loaded/active there.  The only issue is that there is
a slight "double jeopardy" to soaking completion when the ABM is
running repeatedly because the random parameter is re-rolled each
check, so there's a better chance of getting at least one high roll
that completes the soak if you're making repeated rolls.  This is
expected to be a small effect with current recipes.