Fix more junk metadata writes
We weren't resetting the cooking progress correctly for cooking
item stacks (it was using the wrong meta key).

As pointed out by Lars, clearing a metadata key that's marked as
private seems like it triggers a block resend, even if the client
wouldn't see any difference in the mapblock.  This means that
fixing these sorts of issues reduces server-to-client network
traffic, which is (anecdotally) observed with this change on a
copy of the NCC map near spawn.