Experimental replacement for "emergency digging"
One of the most confusing features for new players has been the
fact that they seem to be able to dig anything, but many things
take a long time, and when they finish digging it, they get the
weird "displaced node" that they can't interact with.  It has never
really compensated for this problem by being a compelling way to
get "unstuck" when trapped.

Instead, experiment with a new way to get "untrapped" when trapped
within a solid barrier.  By repeatedly doing the "scaling
placement" recipe on the same spot several times, instead of merely
being able to climb it, players can "phase through" solid matter
and quantum-tunnel themselves through/into a wall.  If there is
space for them on the other side, they will be placed there.  If
not, then they will be shoved into the wall, and can rely on the
"push out of solid" mechanism to eventually wander them over to
somewhere else, where they might have a better chance at getting
out.

This now makes being trapped in nearly arbitrarily deep caves
actually theoretically resolvable, because players should have at
their disposal some combination of navigation inside caves by feel
at their disposal, that lets the make clear progress toward going
"up", and can always try to tunnel into a ceiling for hopes of
conversion to a less deeply buried cave.  This gives players a bit
more agency as well.

Some concern has been expressed about the use of this for bad
actors, but those bad actors have already had more effective ways
to circumvent barriers as it is, which don't suffer from some of
the randomness and unreliability of this method, so the downside
to this is expected to be limited.