Pumice won't fall if it's resting on a liquid
It "floats" on water or pumwater.  Any liquid that has "liquid
movement physics" for players also supports pumice from falling.
This prevents a pumspill over an ocean from filling the ocean
below it with pumice.

This also slightly optmizes the direct-support-below check path
for pumice, as it's now an index lookup instead of looking up the
node definition and running on the fly heuristics.