More exact HUD dot placement on optics
Currently all vanilla rotatable nodes follow a standard of having
a selection box that's a single symmetrical cube centered on the
node position center.  This is already necessary to ensure that
rotating a node doesn't cause the selection box to move out from
under the cursor and cause repeated right-clicks to hit something
else beyond.

Taking advantage of this assumption makes it simple to just add a
single scale factor to:
- Make rotation hit zone regions consistently sized.
- Accurately place the HUD dots indicating the regions.

A more elegant hack may be necessary for nodes that are don't
meet the symmetry requirements, but it may actually just make more
sense to throw warnings if any are registered.