Fix missing toolcaps on hand
Since a current dev Luanti engine change, Luanti is now using
the toolcaps from items in the "hand" inventory slot and NOT
falling back on the "empty item" toolcaps when one is defined.
Copy the default hand toolcaps onto both empty and hand items
to work around this for both old and new versions.